Split up smallCommands into individual commands
