+++ commands/magic.js
@@ -1,0 +1,8 @@
+module.exports = {
+	name: "magic",
+	args: 0,
+	help: "Magic!",
+	execute: (message,args) => {
+		message.channel.send("https://giphy.com/gifs/magical-KFNQuuT1qx7I4");
+	}
+}
