+++ commands/pong.js
@@ -1,0 +1,8 @@
+module.exports = {
+	name: "pong",
+	args: 0,
+	help: "Pong!",
+	execute: (message,args) => {
+		message.react("🏓");
+	}
+}
